# openapi.yaml
openapi: 3.0.3
info:
  title: Olumi Assistants Service API
  version: 1.11.1
  description: |
    AI-powered decision graph drafting service that transforms natural language briefs into structured,
    executable decision models with provenance tracking and validation.

    ## Features
    - **Structured Provenance**: Traceable citations to source documents with locations
    - **Streaming Support**: Server-Sent Events (SSE) for real-time progress updates
    - **LLM-Guided Repair**: Intelligent graph validation and auto-repair
    - **Security Rails**: Rate limiting, body size limits, request timeouts
    - **Deprecation Signals**: Migration support for legacy provenance formats

  contact:
    name: Olumi API Support
    url: https://docs.olumi.ai
    email: support@olumi.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:3101
    description: Local development
  - url: https://api.olumi.ai
    description: Production
...
paths:
...
  /assist/v1/draft-graph:
    post:
      summary: Draft decision graph (CEE v1)
      description: |
        Cognitive Enhancement Engine (CEE) v1 **Draft My Model** endpoint.

        This endpoint drafts a decision graph from a natural language brief using
        the engine's canonical GraphV1 / GraphPatchV1 contracts and the CEE
        validation pipeline. The underlying graph and patch payloads are the
        same engine contracts used by `/assist/draft-graph`; CEE adds trace,
        quality, validation metadata and archetype classification on top.

        **Rate limiting (CEE‑specific):**
        - Default: 5 requests per 60 seconds per API key (or client IP fallback).
        - Configurable via `CEE_DRAFT_RATE_LIMIT_RPM`.
        - Exceeding this per‑feature limit returns `CEEErrorResponseV1` with
          `code="CEE_RATE_LIMIT"`, `retryable=true`, a `Retry-After` header,
          and `details.retry_after_seconds` describing when it is safe to retry.
      tags:
        - Draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CEEDraftGraphRequestV1'
      responses:
        '200':
          description: Graph successfully generated (CEE v1)
          headers:
            X-CEE-API-Version:
              description: CEE API version used to process this request.
              schema:
                type: string
                enum: ["v1"]
            X-CEE-Feature-Version:
              description: Version of the CEE Draft My Model feature.
              schema:
                type: string
            X-CEE-Request-ID:
              description: Unique request identifier for CEE tracing.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CEEDraftGraphResponseV1'
        '400':
          description: CEE validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CEEErrorResponseV1'
        '429':
          description: CEE per-feature rate limit exceeded for Draft My Model
          headers:
            Retry-After:
              description: Seconds until the client may retry this request.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CEEErrorResponseV1'
        '500':
          description: CEE internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CEEErrorResponseV1'
...
components:
  schemas:
...
    DraftGraphOutput:
      type: object
      required:
        - graph
        - patch
      properties:
        graph:
          $ref: '#/components/schemas/Graph'
        patch:
          $ref: '#/components/schemas/GraphPatch'
        rationales:
          type: array
          description: Explanations for graph elements
          items:
            type: object
            properties:
              target:
                type: string
                description: Node or edge ID being explained
              why:
                type: string
                maxLength: 280
                description: Brief explanation
              provenance_source:
                type: string
                description: Optional citation source
        issues:
          type: array
          description: Validation warnings (if any)
          items:
            type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score for the generated graph
          example: 0.87
        clarifier_status:
          type: string
          enum: [complete, max_rounds, confident]
          description: Whether clarification questions are needed
        debug:
          type: object
          description: Debug information (only if include_debug=true)
          properties:
            needle_movers:
              type: array
              description: Document preview metadata
              items:
                type: object
...
    CEEValidationIssue:
      type: object
      description: Validation issue discovered in the CEE pipeline
      required:
        - code
        - severity
      properties:
        code:
          type: string
        severity:
          type: string
          enum: [info, warn, error]
        field:
          type: string
          description: Optional field or path this issue relates to
        message:
          type: string
          description: Human-readable description for clients (not logged as telemetry)
        suggestion:
          type: string
          description: High-level suggestion for how to address this issue
        meta:
          type: object
          description: Additional machine-readable metadata for this issue
          additionalProperties: true
        details:
          type: object
          additionalProperties: true
...
    CEEQualityMeta:
      type: object
      description: Quality scores for CEE outputs (1-10 scale)
      required:
        - overall
      properties:
        overall:
          type: integer
          minimum: 1
          maximum: 10
        structure:
          type: integer
          minimum: 1
          maximum: 10
        structural_proxy:
          type: integer
          minimum: 1
          maximum: 10
        coverage:
          type: integer
          minimum: 1
          maximum: 10
        safety:
          type: integer
          minimum: 1
          maximum: 10
        issues_by_severity:
          type: object
          description: Aggregate counts of validation issues by severity
          properties:
            error:
              type: integer
            warning:
              type: integer
            info:
              type: integer
        details:
          type: object
          additionalProperties: true
...
    CEEErrorResponseV1:
      type: object
      description: Error response schema for CEE endpoints
      required:
        - schema
        - code
        - message
      properties:
        schema:
          type: string
          enum: [cee.error.v1]
        code:
          $ref: '#/components/schemas/CEEErrorCode'
        message:
          type: string
          description: Human-readable error message
        source:
          type: string
          description: Logical source of the error (always "cee" for CEE endpoints)
        request_id:
          type: string
          description: Stable request identifier for tracing and support
        degraded:
          type: boolean
          description: True when CEE or its dependencies are operating in degraded mode
        reason:
          type: string
          description: Short machine-friendly reason code (e.g. "empty_graph", "incomplete_structure")
        recovery:
          type: object
          description: High-level guidance on how to fix input or proceed
          properties:
            suggestion:
              type: string
              description: Single high-level suggestion for the user
            hints:
              type: array
              description: Concrete, short hints for remediation
              items:
                type: string
            example:
              type: string
              description: Optional example brief or graph description
        recovery_suggestion:
          type: string
          description: >-
            Flat mirror of recovery.suggestion — the PINNED producer field
            name consumers read first (wave-2 ask 7, @talchain/schemas
            0.19.0 CeeTypedErrorSchema; routed from DGAI #383, which was
            passthrough-sniffing three fallback names because no flat name
            was typed). Present exactly when recovery.suggestion is.
        node_count:
          type: integer
          description: Graph node count when relevant to the error
        edge_count:
          type: integer
          description: Graph edge count when relevant to the error
        missing_kinds:
          type: array
          description: Node kinds missing from the graph for minimum-structure enforcement
          items:
            type: string
        retryable:
          type: boolean
          description: Whether the client may safely retry this request
        trace:
          $ref: '#/components/schemas/CEETraceMeta'
        details:
          type: object
          additionalProperties: true
    CEEDraftGraphRequestV1:
      allOf:
        - $ref: '#/components/schemas/DraftGraphInput'
        - type: object
          properties:
            seed:
              type: string
              description: Optional client-provided seed for deterministic behaviour
            archetype_hint:
              type: string
              description: Optional hint for the target archetype, e.g. pricing_decision
...
    CEEDraftGraphResponseV1:
      allOf:
        - $ref: '#/components/schemas/DraftGraphOutput'
        - type: object
          required:
            - trace
            - quality
          properties:
            trace:
              $ref: '#/components/schemas/CEETraceMeta'
            quality:
              $ref: '#/components/schemas/CEEQualityMeta'
            validation_issues:
              type: array
              items:
                $ref: '#/components/schemas/CEEValidationIssue'
            archetype:
              type: object
              description: Archetype classification metadata
              properties:
                decision_type:
                  type: string
                  description: Classified decision type, e.g. pricing_decision
                match:
                  type: string
                  enum: [exact, fuzzy, generic]
                confidence:
                  type: number
                  minimum: 0
                  maximum: 1
            seed:
              type: string
              description: Seed value used for this response
            response_hash:
              type: string
              description: Engine-provided response hash for reproducibility
            response_limits:
              type: object
              properties:
                bias_findings_max:
                  type: integer
                bias_findings_truncated:
                  type: boolean
                options_max:
                  type: integer
                options_truncated:
                  type: boolean
                evidence_suggestions_max:
                  type: integer
                evidence_suggestions_truncated:
                  type: boolean
                sensitivity_suggestions_max:
                  type: integer
                sensitivity_suggestions_truncated:
                  type: boolean
            draft_warnings:
              type: array
              description: Optional structural warnings about the draft graph
              items:
                $ref: '#/components/schemas/CEEStructuralWarningV1'
            confidence_flags:
              $ref: '#/components/schemas/CEEConfidenceFlagsV1'
            guidance:
              $ref: '#/components/schemas/CEEGuidanceV1'
            clarifier:
              $ref: '#/components/schemas/CEEClarifierBlockV1'
            weight_suggestions:
              type: array
              description: Suggestions for edge weights that appear uniform/unrefined
              items:
                $ref: '#/components/schemas/CEEWeightSuggestionV1'
            comparison_suggested:
              type: boolean
              description: Whether the graph structure suggests comparison analysis would add value
            goal_connectivity:
              $ref: '#/components/schemas/CEEGoalConnectivityV1'
            model_quality_factors:
              $ref: '#/components/schemas/CEEModelQualityFactorsV1'
            intervention_hints:
              type: array
              description: Metadata hints for interventions to assist normalisation
              items:
                $ref: '#/components/schemas/CEEInterventionHintV1'
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (if enabled)
security:
  - apiKey: []
...